+++ CHANGELOG.md
@@ -5,3 +5,28 @@ Add ZPL text and label.
 ## 0.1.1
 
 Add ZPL network printer.
+
+## 0.1.2
+
+- Change ZPL printer to be sealed
+- Add ZPL element equality
+
+## 0.2.0
+
+- Add barcode element
+
+## 0.2.1
+
+- Add field origin alignment
+
+## 0.2.2
+
+- Add label margins API
+
+## 0.2.3
+
+- Add `ZplPrinter.printAll`
+
+## 0.2.4
+
+- Add `ZplTextStyle.copyWith`